Мелкие доработки
Убрал лишние поля и методы в  классах Enemy и Player

+++ gameASCII/gameASCII/Item.h
@@ -8,14 +8,14 @@ class Item
 {
 public:
 	Item();
-	Item(string name, int price, int updatePrice, float defense, float damage, short level, float defenceUpdateDifference, float damageUpdateDifference);
+	Item(string name, int price, int updatePrice, float defense, int damage, short level, float defenceUpdateDifference, float damageUpdateDifference);
 
 	// Getters
 	string getName()const{ return _name; }
 	int getPrice()const{ return _price; }
 	int getUpdatePrice()const{ return _baseUpdatePrice + BASE_LVL_UPDATE_DIFERENCE * this->_level;; }
 	float getDefense()const{ return _defense + _defenceUpdateDifference * _level; }
-	float getDamage()const{ return _damage + _damageUpdateDifference * _level; }
+	int getDamage()const{ return _damage + _damageUpdateDifference * _level; }
 	short getLevel()const{ return _level; }
 	float getDefenceUpdateDifference()const{ return _defenceUpdateDifference; }
 	float getDamageUpdateDifference()const{ return _damageUpdateDifference; }
@@ -30,6 +30,6 @@ private:
 	float _damageUpdateDifference;
 
 	float _defense;
-	float _damage;
+	int _damage;
 	short _level;
 };
